+++ app/components/layout/cards/feature_component/feature_component.html.erb
