+++ app/components/layout/cards/statement_component/statement_component.html.erb
