+++ app/components/layout/cards/stats_component/stats_component.scss
