+++ app/components/layout/grid_component/grid_component.html.erb
