+++ app/views/home/index.html.erb
@@ -2,7 +2,7 @@
 
 <% if Flipper.enabled?(:new_home_page, current_user) %>
   <div class="bg-blue-very-dark text-white">
-    <%= render GridComponent.new(cols: 2, classes: 'container') do |grid| %>
+    <%= render Layout::GridComponent.new(cols: 2, classes: 'container') do |grid| %>
       <% grid.with_feature_card classes: 'py-4' do |feature| %>
         <%= feature.with_header title: t('home.feature.header'), classes: 'text-white' %>
         <%= feature.with_description { t('home.feature.description_html') } %>
@@ -16,13 +16,13 @@
   </div>
 
   <div class="container">
-    <%= render Cards::StatementComponent.new do |statment| %>
-      <%= statment.with_header title: t('home.statement.header') %>
-      <%= statment.with_description { t('home.statement.description') } %>
+    <%= render Layout::Cards::StatementComponent.new do |statement| %>
+      <%= statement.with_header title: t('home.statement.header') %>
+      <%= statement.with_description { t('home.statement.description') } %>
     <% end %>
 
-    <%= render GridComponent.new(cols: 4,
-                                 cell_classes: 'pb-4') do |grid| %>
+    <%= render Layout::GridComponent.new(cols: 4,
+                                         cell_classes: 'pb-4') do |grid| %>
       <% grid.with_stats_card do |card| %>
         <%= card.with_icon fuel_type: :electricity, style: :circle %>
         <%= card.with_header title: t('home.stats.card_1.header') %>
